+++ openlane/steps/openroad.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Efabless Corporation
+# Copyright 2023-2025 Efabless Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,74 +11,63 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import functools
 import io
+import json
 import os
 import re
-import json
-import textwrap
-import tempfile
-import functools
 import subprocess
-from enum import Enum
-from math import inf
-from glob import glob
-from decimal import Decimal
-from base64 import b64encode
+import tempfile
+import textwrap
+import pathlib
 from abc import abstractmethod
-from dataclasses import dataclass
+from base64 import b64encode
 from concurrent.futures import Future, ThreadPoolExecutor
-from typing import (
-    Any,
-    List,
-    Dict,
-    Literal,
-    Set,
-    Tuple,
-    Optional,
-    Union,
-)
+from dataclasses import dataclass
+from decimal import Decimal
+from enum import Enum
+from glob import glob
+from math import inf
+from typing import Any, Dict, List, Literal, Optional, Set, Tuple, Union
 
-
-import yaml
 import rich
 import rich.table
+import yaml
 
+from ..common import (
+    Path,
+    Filter,
+    TclUtils,
+    DRC as DRCObject,
+    _get_process_limit,
+    aggregate_metrics,
+    get_script_dir,
+    mkdirp,
+    process_list_file,
+)
+from ..config import Macro, Variable
+from ..config.flow import option_variables
+from ..logging import console, debug, info, options, verbose
+from ..state import DesignFormat, State
+from .common_variables import (
+    dpl_variables,
+    grt_variables,
+    io_layer_variables,
+    pdn_variables,
+    routing_layer_variables,
+    rsz_variables,
+)
+from .openroad_alerts import OpenROADAlert, OpenROADOutputProcessor
 from .step import (
     CompositeStep,
     DefaultOutputProcessor,
-    StepError,
-    ViewsUpdate,
     MetricsUpdate,
     Step,
+    StepError,
     StepException,
-)
-from .openroad_alerts import (
-    OpenROADAlert,
-    OpenROADOutputProcessor,
+    ViewsUpdate,
 )
 from .tclstep import TclStep
-from .common_variables import (
-    io_layer_variables,
-    pdn_variables,
-    rsz_variables,
-    dpl_variables,
-    grt_variables,
-    routing_layer_variables,
-)
-
-from ..config import Variable, Macro
-from ..config.flow import option_variables
-from ..state import State, DesignFormat
-from ..logging import debug, info, verbose, console, options
-from ..common import (
-    Path,
-    TclUtils,
-    get_script_dir,
-    mkdirp,
-    aggregate_metrics,
-    process_list_file,
-    _get_process_limit,
-)
 
 EXAMPLE_INPUT = """
 li1 X 0.23 0.46
@@ -204,6 +193,38 @@ class OpenROADStep(TclStep):
 
     config_vars = [
         Variable(
+            "PNR_CORNERS",
+            Optional[List[str]],
+            "A list of fully-qualified IPVT corners to use during PnR. Can be overriden by some steps",
+            pdk=True,
+        ),
+        Variable(
+            "SET_RC_VERBOSE",
+            bool,
+            "If set to true, set_rc commands are echoed. Quite noisy, but may be useful for debugging.",
+            default=False,
+        ),
+        Variable(
+            "LAYERS_RC",
+            Optional[Dict[str, Dict[str, Dict[str, Decimal]]]],
+            "Used during PNR steps, Specific custom resistance and capacitance values for metal layers."
+            + " For each IPVT corner, a mapping for each metal layer is provided."
+            + " Each mapping describes custom resistance and capacitance values."
+            + " Usage of wildcards for specifying IPVT corners is allowed."
+            + " Units are resistance and capacitance per unit length as defined in the first lib file.",
+            pdk=True,
+        ),
+        Variable(
+            "VIAS_R",
+            Optional[Dict[str, Dict[str, Dict[str, Decimal]]]],
+            "Used during PNR steps, Specific custom resistance values for via layers."
+            + " For each IPVT corner, a mapping for each via layer is provided."
+            + " Each mapping describes custom resistance values."
+            + " Usage of wildcards for specifying IPVT corners is allowed."
+            + " Via resistance is per cut/via with units asdefined in the first lib file.",
+            pdk=True,
+        ),
+        Variable(
             "PDN_CONNECT_MACROS_TO_GRID",
             bool,
             "Enables the connection of macros to the top level power grid.",
@@ -274,9 +295,8 @@ class OpenROADStep(TclStep):
         """
         The `run()` override for the OpenROADStep class handles two things:
 
-        1. Before the `super()` call: It creates a version of the lib file
-        minus cells that are known bad (i.e. those that fail DRC) and pass it on
-        in the environment variable `_PNR_LIBS`.
+        1. Before the `super()` call: Process _LIB_CORNER_<i> for liberty/corner
+        pairs.
 
         2. After the `super()` call: Processes the `or_metrics_out.json` file and
         updates the State's `metrics` property with any new metrics in that object.
@@ -284,9 +304,51 @@ class OpenROADStep(TclStep):
         kwargs, env = self.extract_env(kwargs)
         env = self.prepare_env(env, state_in)
 
+        corners: List[str] = self.config["PNR_CORNERS"] or [
+            self.config["DEFAULT_CORNER"]
+        ]
+
+        if "corners" in kwargs:
+            corners = kwargs.pop("corners")
+            debug(f"Corners Override {corners}")
+
+        count = 0
+        for corner in corners:
+            _, libs, _, _ = self.toolbox.get_timing_files_categorized(
+                self.config, corner
+            )
+            env[f"_LIB_CORNER_{count}"] = TclStep.value_to_tcl([corner] + libs)
+            debug(f"Liberty files for '{corner}' added: {libs}")
+            count += 1
+
         check = False
         if "check" in kwargs:
             check = kwargs.pop("check")
+
+        layers_rc = self.config["LAYERS_RC"]
+        if layers_rc is not None:
+            count = 0
+            for corner_wildcard, metal_layers in layers_rc.items():
+                for corner in Filter([corner_wildcard]).filter(corners):
+                    for layer, rc in metal_layers.items():
+                        res = rc["res"]
+                        cap = rc["cap"]
+                        env[f"_LAYER_RC_{count}"] = TclStep.value_to_tcl(
+                            [corner, layer, res, cap]
+                        )
+                        count += 1
+
+        vias_r = self.config["VIAS_R"]
+        if vias_r is not None:
+            count = 0
+            for corner_wildcard, metal_layers in vias_r.items():
+                for corner in Filter(corner_wildcard).filter(corners):
+                    for via, rc in metal_layers.items():
+                        res = rc["res"]
+                        env[f"_VIA_R_{count}"] = TclStep.value_to_tcl(
+                            [corner, via, res]
+                        )
+                        count += 1
 
         command = self.get_command()
 
@@ -913,6 +975,12 @@ class Floorplan(OpenROADStep):
 
     config_vars = OpenROADStep.config_vars + [
         Variable(
+            "FP_FLIP_SITES",
+            Optional[List[str]],
+            "Flip these sites vertically. Useful in niche alignment scenarios where single-height cells have ground at the south side and double-height cells have power at the south side, causing a short. In that situation, flipping the sites for single-height cells resolves the issue.",
+            pdk=True,
+        ),
+        Variable(
             "FP_TRACKS_INFO",
             Path,
             "A path to the a classic OpenROAD `.tracks` file. Used by the floorplanner to generate tracks.",
@@ -1122,6 +1190,13 @@ class TapEndcapInsertion(OpenROADStep):
 
     config_vars = OpenROADStep.config_vars + [
         Variable(
+            "FP_TAPCELL_DIST",
+            Optional[Decimal],
+            "The distance between tap cell columns. Must be specified if WELLTAP_CELL is specified.",
+            units="µm",
+            pdk=True,
+        ),
+        Variable(
             "FP_MACRO_HORIZONTAL_HALO",
             Decimal,
             "Specify the horizontal halo size around macros.",
@@ -1137,17 +1212,18 @@ class TapEndcapInsertion(OpenROADStep):
             units="µm",
             deprecated_names=["FP_TAP_VERTICAL_HALO"],
         ),
-        Variable(
-            "FP_TAPCELL_DIST",
-            Decimal,
-            "The distance between tap cell columns.",
-            units="µm",
-            pdk=True,
-        ),
     ]
 
     def get_script_path(self):
         return os.path.join(get_script_dir(), "openroad", "tapcell.tcl")
+
+    def run(self, state_in, **kwargs):
+        if (
+            self.config["WELLTAP_CELL"] is not None
+            and self.config["FP_TAPCELL_DIST"] is None
+        ):
+            raise StepException("FP_TAPCELL_DIST must be set if WELLTAP_CELL is set.")
+        return super().run(state_in, **kwargs)
 
 
 @Step.factory.register()
@@ -1283,10 +1359,15 @@ class _GlobalPlacement(OpenROADStep):
             ),
             Variable(
                 "GPL_CELL_PADDING",
-                Decimal,
+                int,
                 "Cell padding value (in sites) for global placement. The number will be integer divided by 2 and placed on both sides.",
                 units="sites",
                 pdk=True,
+            ),
+            Variable(
+                "PL_KEEP_RESIZE_BELOW_OVERFLOW",
+                Optional[Decimal],
+                "Only applicable when PL_TIME_DRIVEN is enabled. When the overflow is below the set value, timing-driven iterations will retain the resizer changes instead of reverting them. Allowed values are 0 to 1. If not set, a nonzero default value from OpenROAD will be used",
             ),
         ]
     )
@@ -1603,6 +1684,13 @@ class RepairAntennas(CompositeStep):
 
     Steps = [_DiodeInsertion, CheckAntennas]
 
+    def run(self, state_in: State, **kwargs) -> Tuple[ViewsUpdate, MetricsUpdate]:
+        if self.config["DIODE_CELL"] is None:
+            info(f"'DIODE_CELL' not set. Skipping '{self.id}'…")
+            return {}, {}
+
+        return super().run(state_in, **kwargs)
+
 
 @Step.factory.register()
 class DetailedRouting(OpenROADStep):
@@ -1652,9 +1740,22 @@ class DetailedRouting(OpenROADStep):
             default=False,
         ),
         Variable(
+            "DRT_ANTENNA_REPAIR_ITERS",
+            int,
+            "The maximum number of iterations to run antenna repair. Set to a positive integer to attempt to repair antennas and then re-run DRT as appropriate.",
+            default=0,
+        ),
+        Variable(
+            "DRT_ANTENNA_MARGIN",
+            int,
+            "The margin to over fix antenna violations.",
+            default=10,
+            units="%",
+        ),
+        Variable(
             "DRT_SAVE_DRC_REPORT_ITERS",
             Optional[int],
-            "Report DRC on each specified iteration. Set to 1 when DRT_SAVE_DRC_REPORT_ITERS in enabled",
+            "Write a DRC report every N iterations. If DRT_SAVE_SNAPSHOTS is enabled, there is an implicit default value of 1.",
         ),
     ]
 
@@ -1665,7 +1766,21 @@ class DetailedRouting(OpenROADStep):
         kwargs, env = self.extract_env(kwargs)
         env["DRT_THREADS"] = env.get("DRT_THREADS", str(_get_process_limit()))
         info(f"Running TritonRoute with {env['DRT_THREADS']} threads…")
-        return super().run(state_in, env=env, **kwargs)
+        views_updates, metrics_updates = super().run(state_in, env=env, **kwargs)
+
+        drc_paths = list(pathlib.Path(self.step_dir).rglob("*.drc*"))
+        for path in drc_paths:
+            drc, _ = DRCObject.from_openroad(
+                open(path, encoding="utf8"), self.config["DESIGN_NAME"]
+            )
+
+            drc.to_klayout_xml(open(pathlib.Path(str(path) + ".xml"), "wb"))
+        #        if violation_count > 0:
+        #            self.warn(
+        #                f"DRC errors found after routing. View the report file at {report_path}.\nView KLayout xml file at {klayout_db_path}"
+        #            )
+
+        return views_updates, metrics_updates
 
 
 @Step.factory.register()
@@ -2025,33 +2140,16 @@ class ResizerStep(OpenROADStep):
         **kwargs,
     ) -> Tuple[ViewsUpdate, MetricsUpdate]:
         kwargs, env = self.extract_env(kwargs)
-
-        corners_key: str = "RSZ_CORNERS"
-
-        if "corners_key" in kwargs:
-            corners_key = kwargs.pop("corners_key")
-
-        corners = self.config[corners_key] or self.config["STA_CORNERS"]
-        lib_set_set = set()
-        count = 0
-        for corner in corners:
-            _, libs, _, _ = self.toolbox.get_timing_files_categorized(
-                self.config, corner
-            )
-            lib_set = frozenset(libs)
-            if lib_set in lib_set_set:
-                debug(f"Liberty files for '{corner}' already accounted for- skipped")
-                continue
-            lib_set_set.add(lib_set)
-            env[f"RSZ_CORNER_{count}"] = TclStep.value_to_tcl([corner] + libs)
-            debug(f"Liberty files for '{corner}' added: {libs}")
-            count += 1
-
-        return super().run(state_in, env=env, **kwargs)
+        return super().run(
+            state_in,
+            corners=self.config["RSZ_CORNERS"] or self.config["STA_CORNERS"],
+            env=env,
+            **kwargs,
+        )
 
 
 @Step.factory.register()
-class CTS(ResizerStep):
+class CTS(OpenROADStep):
     """
     Creates a `Clock tree <https://en.wikipedia.org/wiki/Clock_signal#Distribution>`_
     for an ODB file with detailed-placed cells, using reasonably accurate resistance
@@ -2139,7 +2237,7 @@ class CTS(ResizerStep):
             Variable(
                 "CTS_CLK_BUFFERS",
                 List[str],
-                "Defines the list of clock buffers to be used in CTS.",
+                "Defines the list of clock buffer names or buffer name wildcards to be used in CTS.",
                 deprecated_names=["CTS_CLK_BUFFER_LIST"],
                 pdk=True,
             ),
@@ -2176,7 +2274,10 @@ class CTS(ResizerStep):
                 return {}, {}
 
         views_updates, metrics_updates = super().run(
-            state_in, corners_key="CTS_CORNERS", env=env, **kwargs
+            state_in,
+            corners=self.config["CTS_CORNERS"] or self.config["STA_CORNERS"],
+            env=env,
+            **kwargs,
         )
 
         return views_updates, metrics_updates
@@ -2580,3 +2681,22 @@ class OpenGUI(Step):
             )
 
         return {}, {}
+
+
+@Step.factory.register()
+class DumpRCValues(OpenROADStep):
+    """
+    Creates three reports:
+
+    * Initial Database Layer RC Values (from Tech LEF)
+    * Modified Database Layer RC Values
+    * Modified Resizer Layer RC Values
+    """
+
+    id = "OpenROAD.DumpRCValues"
+    name = "Dump RC Values"
+
+    inputs = [DesignFormat.DEF]
+
+    def get_script_path(self) -> str:
+        return os.path.join(get_script_dir(), "openroad", "dump_rc.tcl")