+++ Changelog.md
@@ -28,22 +28,41 @@
   * `PDN_OBSTRUCTIONS` and `ROUTING_OBSTRUCTIONS` are now lists of tuples
     instead of variable-length Tcl-style lists (AKA: strings).
 
+* `Odb.DiodesOnPorts`, `Odb.PortDiodePlacement`
+  * Steps no longer assume `DIODE_CELL` exists and fall back to doing nothing.
+
+*  `Odb.FuzzyDiodePlacement`, `Odb.HeuristicDiodeInsertion`
+  * Steps no longer assume `DIODE_CELL` exists and fall back to doing nothing.
+  * `HEURISTIC_ANTENNA_THRESHOLD` has been made optional, steps do nothing if
+    it is unset.
+
 * `OpenROAD.*`
 
+  * Added `PNR_CORNERS`. An override for `DEFAULT_CORNER` for PnR steps except
+    for steps using `RSZ_CORNERS` and `CTS_CORNERS`.
+  * Added `LAYERS_RC`, `VIAS_R`: Unlike OpenLane 1.0.0 variables with similar
+    names, these are mappings from corners to layer/via RC values.
+    * `PNR_CORNERS`, `RSZ_CORNERS`, and `CTS_CORNERS` all now support multiple
+      corners to have the same set of liberty files (as RC values may differ.)
+  * Added `SET_RC_VERBOSE`, which (very noisily) logs set-RC-related commands to
+    logs.
+  * Always read libs before reading odb.
   * Added `log_cmd` from OpenROAD-flow-scripts -- neat idea for consistency
-  * New convenience methods to append flags to calls based on environment
-    variables
+  * Lib files are now *always* read BEFORE reading database files.
   * **Internal**: Steps now sensitive to `_OPENROAD_GUI` environment variable --
     coupled with `--only`, it runs a step in OpenROAD then doesn't quit so you
     may inspect the result.
     * This is not part of the OpenLane stable API and may be broken at any
       moment.
+  * **Internal**: New convenience methods to append flags to calls based on
+    environment variables
 
 * `OpenROAD.CTS`
 
   * Added flags `CTS_OBSTRUCTION_AWARE` and `CTS_BALANCE_LEVELS`
   * Added `CTS_SINK_BUFFER_MAX_CAP_DERATE_PCT`
   * Added `CTS_DELAY_BUFFER_DERATE_PCT`
+  * `CTS_CLK_BUFFERS` can now take wildcards.
 
 * `OpenROAD.CutRows`
 
@@ -55,10 +74,35 @@
   * Added `DRT_SAVE_SNAPSHOTS` which enables saving snapshots of the layout each
     detalied routing iteration.
   * Added `DRT_SAVE_DRC_REPORT_ITERS`
+  * Added `DRT_ANTENNA_REPAIR_ITERS`, which, if greater than zero and
+    `DIODE_CELL` is set, enables antenna fixing after detailed routing
+  * Added `DRT_ANTENNA_MARGIN` which is similar to `GRT_ANTENNA_MARGIN` but for
+    the aforementioned antenna repair iterations
+  * DRC reports are now converted to `xml` and readable by KLayout
+
+* Created `OpenROAD.DumpRCValues`
+
+  * Creates three reports to help verify that the RC values used for estimation
+    are set correctly.
+
+* `OpenROAD.Floorplan`
+
+  * Added `FP_FLIP_SITES`: allows sites in floorplans to be flipped. Useful in
+    niche alignment scenarios where single-height cells have ground at the south
+    side and double-height cells have power at the south side, causing a short.
+    In that situation, flipping the sites for single-height cells resolves the
+    issue.
 
 * `OpenROAD.GlobalPlacement`
 
   * Added optional variable `PL_ROUTABILITY_MAX_DENSITY_PCT`
+  * Added optional variable `PL_KEEP_RESIZE_BELOW_OVERFLOW`
+
+  * Corrected `GPL_CELL_PADDING` to be an integer.
+
+* `OpenROAD.RepairAntennas`
+
+  * Step no longer assumes `DIODE_CELL` exists and falls back to doing nothing.
 
 * `OpenROAD.RepairDesignPostGPL`
 
@@ -94,6 +138,11 @@
     * `GRT_RESIZER_HOLD_REPAIR_TNS_PCT`
     * `GRT_RESIZER_HOLD_MAX_UTIL_PCT`
 
+* `OpenROAD.TapDecapInsertion`
+
+  * No longer assumes `WELLTAP_CELL` has a value and skips tap insertion if not.
+  * No longer assumes `DECAP_CELL` has a value and skips decap insertion if not.
+
 * Created `OpenROAD.UnplaceAll`
 
   * Removes the placement status of all instances.
@@ -101,6 +150,11 @@
 * `Yosys.*Synthesis`
 
   * Added `SYNTH_CORNER`: a step-specific override for `DEFAULT_CORNER`.
+
+## Flows
+
+* Classic
+  * Added `OpenROAD.DumpRCValues` immediately after floorplanning.
 
 ## Tool Updates
 
@@ -122,6 +176,24 @@
 
 ## Misc. Enhancements/Bugfixes
 
+* `openlane.flows`
+
+  * `SequentialFlow`
+    * Substitutions are now to be strictly consumed by the subclass initializer,
+      i.e., it can no longer be done on the object-level and only on the class
+      level. Additionally, it can provided as a list of tuples instead of a
+      dictionary so the same key may be reused multiple times.
+    * Step IDs are re-normalized after every substitution, so a substitution for
+      `OpenROAD.DetailedPlacement-1` for example would always refer to the
+      second `OpenROAD.DetailedPlacement` AFTER applying all previous
+      substitutions, instead of the second "original"
+      `OpenROAD.DetailedPlacement` in the flow.
+
+* `openlane.config`
+
+  * `meta.substituting_steps` now only apply to the sequential flow declared in
+    `meta.flow` and not all flows.
+
 * `openlane.state`
 
   * `DesignFormat`
@@ -139,6 +211,11 @@
     * States initialized with keys that have values that are `None` now remove
       said keys.
 
+* `openlane.steps`
+
+  * TclStep
+    * All `Decimal` values are now passed to Tcl in exponent notation.
+
 * `openlane.config`
 
   * Moved a number of global variables:
@@ -155,6 +232,12 @@
 
 ## API Breaks
 
+* `*`
+
+  * `{GPL,DPL}_CELL_PADDING`, `PL_MAX_DISPLACEMENT_{X,Y}` now all integers to
+    match OpenROAD.
+  * `WELLTAP_CELL`, `DECAP_CELL` now optional.
+
 * `Checker.HoldViolations`
 
   * `HOLD_VIOLATION_CORNERS` now defaulting to all corners will require designs
@@ -166,6 +249,20 @@
   * Typing for representation of obstructions has been changed. Designs with a
     meta version of 2 or higher must update their variables from strings to
     tuples.
+
+* `OpenROAD.*`
+
+  * `LAYERS_RC` now uses a new format. Refer to the documentation for a
+    description of the new format.
+  * `VIAS_RC` removed and replaced by `VIAS_R` with a format similar to
+    `LAYERS_RC`.
+
+* `openlane.flows`
+
+  * Step IDs are re-normalized after every substitution, so a substitution for
+    `OpenROAD.DetailedPlacement-1` for example would always refer to the second
+    `OpenROAD.DetailedPlacement` AFTER applying all previous substitutions,
+    instead of the second "original" `OpenROAD.DetailedPlacement` in the flow.
 
 * `openlane.steps`
 
@@ -186,11 +283,29 @@
 
 * `openlane.config`
 
+  * `meta.substituting_steps` now only apply to the sequential flow declared in
+    `meta.flow` and not all flows.
+
   * `WIRE_LENGTH_THRESHOLD`, `GPIO_PAD_*`, `FP_TRACKS_INFO`, `FP_TAPCELL_DIST`
     are no longer global variables.
 
   * `FILL_CELL`, `DECAP_CELL`, `EXTRA_GDS_FILES`, `FALLBACK_SDC_FILE` were all
     renamed, see Misc. Enhancements/Bugfixes.
+
+* `openlane.common.drc`
+
+  * `BoundingBox` changed from `Tuple` to `dataclass` with additional optional
+    `info` property.
+
+# 2.3.3
+
+## Steps
+
+* `OpenROAD.Floorplan`
+
+  * Fixed an issue in `FP_SIZING`: `absolute` mode where if the die area's x0 >
+    x1 or y0 > y1, the computed core area would no longer fit in the die area.
+    Not that we recommend you ever do that, but technically OpenROAD allows it.
 
 # 2.3.2